ChunkReader を Ssq.IO に移動
