+++ Ssq.AnalyzeTool/Program.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Ssq.Printing;
+using Ssq.IO;
 
 namespace Ssq.AnalyzeTool
 {
